Fixed a bug where stars could only be whole.

+++ ideas/style/template/star.js
@@ -1,7 +1,7 @@
 $('.rating').each(function() {
 	var $this = $(this),
 		url = $this.data('vote'),
-		width = parseInt($this.text()) * 25;
+		width = parseFloat($this.text()) * 25;
 
 	$this.html('<ul class="star-rating' + (url ? ' active' : '') + '"><li class="current-rating" style="width: ' + width + 'px"></li><li><a class="one-star">1</a></li><li><a class="two-stars">2</a></li><li><a class="three-stars">3</a></li><li><a class="four-stars">4</a></li><li><a class="five-stars">5</a></li></ul>');
 	$this.find('a').click(function(e) {
